Fix Qt 5.15 deprecation warnings

+++ src/Array2D.hpp
@@ -115,13 +115,13 @@ public:
     }
     void fillBorders( T val ) {
         if(dy > 0) {
-            for(size_t i = 0; i < dy; ++i)
+            for(size_t i = 0; i < static_cast<size_t>(dy); ++i)
                 for(size_t j = 0; j < width; ++j)
                     data[i*width + j] = val;
         }
         if(dx > 0) {
             for(size_t i = 0; i < height; ++i)
-                for(size_t j = 0; j < dx; ++j)
+                for(size_t j = 0; j < static_cast<size_t>(dx); ++j)
                     data[i*width + j] = val;
         }
     }
